+++ app/dfda/page.tsx
@@ -1,7 +1,4 @@
-"use client"
-
 import React from "react"
-
 import DFDAHomePage from "./components/dfda-home-page"
 
 export default function DFDAPage() {
